// types/UserType.d.ts
import { HistoryType } from "./HistoryType"

export interface UserType{
    id?: string,
    berat?: number,
    jk?: string,
    nama?: string,
    pekerjaan?: string,
    penyakit?: string,
    role?: string,
    status?: string,
    telepon?: string,
    tinggi?: number,
    usia?: string,
    email?: string
}